+++ src/main/java/com/example/recipes/service/IngredientService.java
@@ -1,53 +1,18 @@
 package com.example.recipes.service;
 
 import com.example.recipes.model.Ingredient;
-import org.springframework.stereotype.Service;
 
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-@Service
-public class IngredientService {
-    private final Map<Integer, Ingredient> ingredients = new HashMap<>();
 
-    public Ingredient addIngredient(Ingredient ingredient) {
-        if (ingredients.containsKey(ingredient.getId())) {
-            throw new RuntimeException("This ingredient already exists");
-        } else {
-            ingredients.put(ingredient.getId(), ingredient);
-        }
-        return ingredient;
-    }
+public interface IngredientService {
 
-    public Ingredient updateIngredient(int id, Ingredient ingredient) {
-        Ingredient serviceIngredient = ingredients.get(id);
-        if (serviceIngredient == null) {
-            throw new RuntimeException("No such ingredient");
-        }
-        serviceIngredient.setIngredientName(ingredient.getIngredientName());
-        serviceIngredient.setQuantityOfIngredients(ingredient.getQuantityOfIngredients());
-        serviceIngredient.setUnit(ingredient.getUnit());
-        return serviceIngredient;
-    }
+    Ingredient addIngredient(Ingredient ingredient);
 
-    public Ingredient deleteIngredient(int id) {
-        if (ingredients.containsKey(id)) {
-            return ingredients.remove(id);
-        } else {
-            throw new RuntimeException("No such ingredient");
-        }
-    }
+    Ingredient updateIngredient(int id, Ingredient ingredient);
 
-    public Ingredient getIngredient(int id) {
-        if (ingredients.containsKey(id)) {
-            return ingredients.get(id);
-        } else {
-            throw new RuntimeException("No such ingredient");
-        }
-    }
+    Ingredient deleteIngredient(int id);
 
-    public Collection<Ingredient> getAllIngredients() {
-        return ingredients.values();
-    }
+    Ingredient getIngredient(int id);
 
+    Collection<Ingredient> getAllIngredients();
 }
